% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{aparticle_dependency}
\alias{aparticle_dependency}
\title{Dependency}
\usage{
aparticle_dependency(cdn = FALSE)
}
\arguments{
\item{cdn}{Whether to use local or CDN files.}
}
\description{
Add particle dependency.
}
\note{
Source \link[aframer]{a_dependency} \emph{before} \code{aparticle_dependency}.
}
\examples{
library(aframer)

browse_aframe(
  a_scene(
    a_dependency(cdn = TRUE),
    aparticle_dependency(cdn = TRUE),
    a_sky(color = "#606060"),
    a_entity(
      position = xyz_aframe(0, 2.25, -15),
      `particle-system` = opts_aframe(color = "#EF0000,#44CC00")
    )
  )
)

}
